Always hide zone confirmation at start of zone.
* Move zone confirmation logic to questManager module and hide zone
confirmation when zone deck is shown. This was the simplest solution to
ensuring that anytime the showDeck method (walking into a zone, clicking
on a completed zone, and clicking on a zone in the "My Quests" list) is
called, the zone confirmation is hidden.

Refs #142, #153
