+++ src/js/questManager.js
@@ -4,9 +4,11 @@ var _ = require('lodash'),
     $ = require('jquery'),
     Bacon = require('baconjs'),
     BootstrapDialog = require('bootstrap-dialog'),
+    path = require('path'),
     cards = require('./cards'),
     zoneUtils = require('./zoneUtils'),
-    quests = require('../quests.json');
+    quests = require('../quests.json'),
+    zoneConfirmationTemplate = require('../templates/zoneConfirmation.ejs');
 
 var ACTIVE_QUEST = 'fairmount-water-works',
     DEFAULT_PRIMARY_CAPTION = 'Align your screen so that it matches up with the view you see here, tap to show/hide captions.',
@@ -26,6 +28,7 @@ module.exports = {
                 .doAction(onZoneFinished);
 
         initStatus();
+        initZoneConfirmation(zoneChangeStream);
 
         zones.forEach(function(z) {
             var firstPhoto = _.first(z.primaryItems);
@@ -121,5 +124,77 @@ function closeBootstrapDialog(dialog) {
 }
 
 function showDeck(zone) {
+    hideZoneConfirmation();
     cards.openZoneDeck(zone, ACTIVE_QUEST);
+}
+
+function initZoneConfirmation(zoneChangeStream) {
+    zoneChangeStream
+        .doAction(hideZoneConfirmation)
+        .filter(_.isObject)
+        .doAction(updateDialog)
+        .onValue(activateZoneConfirmation);
+}
+
+function updateDialog(zone) {
+    var $dialog = $('#enterarea'),
+        context = {
+            zone: zone,
+            mediaPath: getConfirmationMediaPath(zone)
+        };
+
+    $dialog.html(zoneConfirmationTemplate(context));
+
+    bindConfirmationButtonEvents($dialog, zone);
+}
+
+function getConfirmationMediaPath(zone) {
+    var basePath = path.join('quests', ACTIVE_QUEST, zone.id),
+        mediaPath;
+
+    if (zone.primaryItems[1].type && zone.primaryItems[1].type === 'video') {
+        mediaPath = path.join(basePath, 'media', zone.primaryItems[1].poster);
+    } else {
+        mediaPath = path.join(basePath, 'primary', zone.primaryItems[1].name + '.jpg');
+    }
+
+    return mediaPath;
+}
+
+function bindConfirmationButtonEvents($dialog, zone) {
+    $dialog
+        .find('.enterarea-button.enter')
+        .click(function() {
+            showDeck(zone);
+        });
+
+    $dialog
+        .find('.enterarea-button.dismiss')
+        .click(function() {
+            dismissZoneConfirmation();
+        });
+}
+
+function activateZoneConfirmation(zone) {
+    if (zone.status === zoneUtils.STATUS_FINISHED) {
+        // If the zone has already been completed, only show the small
+        // zone confirmation window.
+        $('#enterarea').addClass('enterarea-dismissed active');
+    } else {
+        $('#enterarea').removeClass('enterarea-dismissed').addClass('active');
+    }
+}
+
+function dismissZoneConfirmation() {
+    $('#enterarea').addClass('enterarea-dismissed');
+}
+
+function hideZoneConfirmation() {
+    var $dialog = $('#enterarea');
+
+    if ($dialog.hasClass('enterarea-dismissed')) {
+        $dialog.removeClass('active');
+    } else {
+        $dialog.removeClass('active enterarea-dismissed');
+    }
 }
